requirements.txt
tensorflow==1.7
scipy
scikit-learn
opencv-python
h5py
matplotlib
Pillow
psutil
cvxpy
imageio
certifi==2019.3.9
chardet==3.0.4
click==6.7
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
requests==2.21.0
SQLAlchemy==1.3.1
urllib3==1.24.1
Werkzeug==0.14.1
pycparser==2.19
six==1.11.0
bcrypt==3.1.4
cffi==1.11.5
